Issue #39 - Define hash for "nocase" strings
Define a hash function for case-insensitive strings. Otherwise the get() function will not consistently find the proper values when given keys that are not of the same case as the initial input to the dictionary. For example a key of "td" may not have been found for an entry that was defined using the key "TD".