ARI updates for object references in text form (#52)
* Issue #39 - Define hash for "nocase" strings

Define a hash function for case-insensitive strings. Otherwise the get() function will not consistently find the proper values when given keys that are not of the same case as the initial input to the dictionary. For example a key of "td" may not have been found for an entry that was defined using the key "TD".

* Issue #39 - Add tests for mixed-case objref types

* Issue #39 - First cut at typechecking objrefs

* Remove dead code

* Issue #39 - Final set of objref updates

Two main changes here:

- Type check objpaths when decoding CBOR
- Update unit tests to pass, including previous tests that were using invalid objpath types

* Issue #39 - Mark new function as local-only