SBR rule execution (#122)
* Add a TODO for potential new CTRL's for rules

* WIP

* Refining TBR execution implementation

* Cleanup, adding TODO's

* Clarify TBR exec_count tracks execs since rule was enabled

* Use ARI access functions from access.c

* Added absolute time to rule, to handle relative start times

This simplifies the logic within exec.c, and adds an exercise for the initializer when creating a rule.

* Add more details for new rule CTRL's

* Rename primary TBR start function, handle state init

Rename to refda_exec_tbr_enable and handle start-up book keeping for the rule.

* Simplify refda_exec_tbr_action

Use existing function to execute the tbr->action macro

* Extent timeline to support execution of TBR

Allow TBR's to be executed via the timeline as well as CTRL sequences.

TBR is simpler since it is just a function executed at a scheduled time, there is not the extra complication of execution sequences that are required for CTRL's.

* Schedule next TBR call after its action is executed

* Initialize abs start time

* Break out worker thread execution

By moving this to its own function we can run a worker thread iteration from our unit tests.

Also ensure context pointer initialization for CTRL's executed from a TBR.

* Add test case for TBR's

* Apply formatting

* Added refda_amm_tbr_desc_reached_max_exec_count

Centralize this logic in tbr.c

* Clean up resources

* Explicitly cancel threads prior to joining

This prevents cases where a thread is unresponsive, interrupting shutdown

* Code review feedback

* Remove pthread_cancel

* Fixes from code review

* Validate TBR action prior to enabling rule

Also updated doxygen comment to clarify

* Cleanup

* Renamed to REFDA_TIMELINE_EXEC

This makes the enum consistent with the union, as well as timeline_exec types.

* WIP

* Building out SBR execution

* Add unit test stub for SBR

* Cleanup of SBR unit test

* Further build-out of SBR conditions

* Formatting, add more SBR test cases

* Improved logging

* Rule improvements

- Update agent performance counters
- Disable rules after max_exec_count is reached

* Add rule disable functions

* Extend tbr tests

- Support absolute start time
- Support multiple iterations

* Cleanup

* Add guards

* Schedule SBR independently of condition complexity

* Rename functions per code review

Make it clearer that running an SBR does not necessary imply anything will be executed.

* Fix compilation error for CHKFALSE

* Schedule next TBR independently of rule expansion

* Improve SBR unit tests

- Expand original SBR test cases
- Add an EDD and test SBR condition false, update EDD state, verify SBR condition becomes true

* Switch EDD backing variable to be an atomic

* Change var to int